Move file related constants to file.ts adn some clean up
